refactor: Use hooks for menu items
