+++ src/components/Menu/config/config.ts
@@ -10,158 +10,130 @@ import {
   NftIcon,
   NftFillIcon,
   MoreIcon,
-  menuStatus,
 } from '@pancakeswap/uikit'
 import { ContextApi } from 'contexts/Localization/types'
 import { nftsBaseUrl } from 'views/Nft/market/constants'
 
 export type ConfigMenuItemsType = MenuItemsType & { hideSubNav?: boolean }
 
-const config: (t: ContextApi['t'], menuItemsStatus?: Record<string, string>) => ConfigMenuItemsType[] = (
-  t,
-  menuItemsStatus,
-) => {
-  const menuItems = [
-    {
-      label: t('Trade'),
-      icon: SwapIcon,
-      fillIcon: SwapFillIcon,
-      href: '/swap',
-      showItemsOnMobile: false,
-      items: [
-        {
-          label: t('Exchange'),
-          href: '/swap',
-        },
-        {
-          label: t('Liquidity'),
-          href: '/liquidity',
-        },
-        {
-          label: t('Limit Orders'),
-          href: '/limit-orders',
-        },
-      ],
-    },
-    {
-      label: t('Earn'),
-      href: '/farms',
-      icon: EarnIcon,
-      fillIcon: EarnFillIcon,
-      items: [
-        {
-          label: t('Farms'),
-          href: '/farms',
-        },
-        {
-          label: t('Pools'),
-          href: '/pools',
-        },
-      ],
-    },
-    {
-      label: t('Win'),
-      href: '/prediction',
-      icon: TrophyIcon,
-      fillIcon: TrophyFillIcon,
-      items: [
-        {
-          label: t('Trading Competition'),
-          href: '/competition',
-        },
-        {
-          label: t('Prediction (BETA)'),
-          href: '/prediction',
-        },
-        {
-          label: t('Lottery'),
-          href: '/lottery',
-        },
-      ],
-    },
-    {
-      label: t('NFT'),
-      href: `${nftsBaseUrl}`,
-      icon: NftIcon,
-      fillIcon: NftFillIcon,
-      items: [
-        {
-          label: t('Overview'),
-          href: `${nftsBaseUrl}`,
-        },
-        {
-          label: t('Collections'),
-          href: `${nftsBaseUrl}/collections`,
-        },
-        {
-          label: t('Activity'),
-          href: `${nftsBaseUrl}/activity`,
-        },
-      ],
-    },
-    {
-      label: '',
-      href: '/info',
-      icon: MoreIcon,
-      hideSubNav: true,
-      items: [
-        {
-          label: t('Info'),
-          href: '/info',
-        },
-        {
-          label: t('IFO'),
-          href: '/ifo',
-        },
-        {
-          label: t('Voting'),
-          href: '/voting',
-        },
-        {
-          type: DropdownMenuItemType.DIVIDER,
-        },
-        {
-          label: t('Leaderboard'),
-          href: '/teams',
-        },
-        {
-          type: DropdownMenuItemType.DIVIDER,
-        },
-        {
-          label: t('Blog'),
-          href: 'https://medium.com/pancakeswap',
-          type: DropdownMenuItemType.EXTERNAL_LINK,
-        },
-        {
-          label: t('Docs'),
-          href: 'https://docs.pancakeswap.finance',
-          type: DropdownMenuItemType.EXTERNAL_LINK,
-        },
-      ],
-    },
-  ]
-
-  if (menuItemsStatus && Object.keys(menuItemsStatus).length) {
-    return menuItems.map((item) => {
-      const innerItems = item.items.map((innerItem) => {
-        const itemStatus = menuItemsStatus[innerItem.href]
-        if (itemStatus) {
-          let itemMenuStatus
-          if (itemStatus === 'soon') {
-            itemMenuStatus = menuStatus.SOON
-          } else if (itemStatus === 'live') {
-            itemMenuStatus = menuStatus.LIVE
-          } else {
-            itemMenuStatus = menuStatus.NEW
-          }
-          return { ...innerItem, status: itemMenuStatus }
-        }
-        return innerItem
-      })
-      return { ...item, items: innerItems }
-    })
-  }
-  return menuItems
-}
+const config: (t: ContextApi['t']) => ConfigMenuItemsType[] = (t) => [
+  {
+    label: t('Trade'),
+    icon: SwapIcon,
+    fillIcon: SwapFillIcon,
+    href: '/swap',
+    showItemsOnMobile: false,
+    items: [
+      {
+        label: t('Exchange'),
+        href: '/swap',
+      },
+      {
+        label: t('Liquidity'),
+        href: '/liquidity',
+      },
+      {
+        label: t('Limit Orders'),
+        href: '/limit-orders',
+      },
+    ],
+  },
+  {
+    label: t('Earn'),
+    href: '/farms',
+    icon: EarnIcon,
+    fillIcon: EarnFillIcon,
+    items: [
+      {
+        label: t('Farms'),
+        href: '/farms',
+      },
+      {
+        label: t('Pools'),
+        href: '/pools',
+      },
+    ],
+  },
+  {
+    label: t('Win'),
+    href: '/prediction',
+    icon: TrophyIcon,
+    fillIcon: TrophyFillIcon,
+    items: [
+      {
+        label: t('Trading Competition'),
+        href: '/competition',
+      },
+      {
+        label: t('Prediction (BETA)'),
+        href: '/prediction',
+      },
+      {
+        label: t('Lottery'),
+        href: '/lottery',
+      },
+    ],
+  },
+  {
+    label: t('NFT'),
+    href: `${nftsBaseUrl}`,
+    icon: NftIcon,
+    fillIcon: NftFillIcon,
+    items: [
+      {
+        label: t('Overview'),
+        href: `${nftsBaseUrl}`,
+      },
+      {
+        label: t('Collections'),
+        href: `${nftsBaseUrl}/collections`,
+      },
+      {
+        label: t('Activity'),
+        href: `${nftsBaseUrl}/activity`,
+      },
+    ],
+  },
+  {
+    label: '',
+    href: '/info',
+    icon: MoreIcon,
+    hideSubNav: true,
+    items: [
+      {
+        label: t('Info'),
+        href: '/info',
+      },
+      {
+        label: t('IFO'),
+        href: '/ifo',
+      },
+      {
+        label: t('Voting'),
+        href: '/voting',
+      },
+      {
+        type: DropdownMenuItemType.DIVIDER,
+      },
+      {
+        label: t('Leaderboard'),
+        href: '/teams',
+      },
+      {
+        type: DropdownMenuItemType.DIVIDER,
+      },
+      {
+        label: t('Blog'),
+        href: 'https://medium.com/pancakeswap',
+        type: DropdownMenuItemType.EXTERNAL_LINK,
+      },
+      {
+        label: t('Docs'),
+        href: 'https://docs.pancakeswap.finance',
+        type: DropdownMenuItemType.EXTERNAL_LINK,
+      },
+    ],
+  },
+]
 
 export default config
